Check arguments of public API
Arguments of the public API were not checked if they are valid.
It was for example possible to create a Mock object for an empty
namespace. The public API will now check its arguments and eventually
bail out with an InvalidArgumentException.

This change is not considered as a backward incompatible change, as
the previous behaviour is considered as a bug. Therefore this commit
will only affect the patch version of the next release.