Record spy invocation in case of an exception
Previously an exception in a spied function would not be recorded as
invocation. This commit will record the invocations arguments and null
as return value. Additionally two methods were added:
Invocation::isExceptionThrown() and Invocation::getException().

This commit changes the behaviour of Spy. Therefore this commit
increases the major version of the next release.